refactor(jotster): checkpoint js-surface migration state

+++ packages/uploads/src/repo/create-attachment.ts
@@ -8,7 +8,7 @@ interface CreateAttachmentInput {
   userId: string;
   fileName: string;
   pathId: string;
-  size: int;
+  size: long;
   contentType: string;
 }
